// src/database/types.ts
export interface FavoriteEntity {
  id?: number;
  productId: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  thumbnail: string;
}

export interface CartEntity {
  id?: number;
  productId: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  thumbnail: string;
  quantity?: number;
}